Add semantic token struct and serialization (#11)
Semantic tokes are encoded as an array of UInt32 in the LSP spec, which makes sense for transport efficiency. But this is not a convenient representation in the LSP server.

This changeset adds a `SemanticToken` representation, and encode/decode functionality to the transport representation.

It also sort the tokens during encode since the token encoding is offset based with unsigned integers, and must be ordered properly.